+++ laokou-common/laokou-common-core/src/main/java/org/laokou/common/core/utils/TemplateUtil.java
@@ -20,6 +20,8 @@ package org.laokou.common.core.utils;
 import freemarker.template.Configuration;
 import freemarker.template.Template;
 import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
+import org.laokou.common.i18n.common.exception.SystemException;
 import org.springframework.ui.freemarker.FreeMarkerTemplateUtils;
 
 import java.io.IOException;
@@ -32,6 +34,7 @@ import static freemarker.template.Configuration.DEFAULT_INCOMPATIBLE_IMPROVEMENT
  *
  * @author laokou
  */
+@Slf4j
 public final class TemplateUtil extends FreeMarkerTemplateUtils {
 
 	/**
@@ -45,10 +48,14 @@ public final class TemplateUtil extends FreeMarkerTemplateUtils {
 	 * @param params 参数
 	 * @return 内容
 	 */
-	@SneakyThrows
 	public static String getContent(String template, Map<String, Object> params) {
-		Template temp = getTemplate(template);
-		return FreeMarkerTemplateUtils.processTemplateIntoString(temp, params);
+		try {
+			Template temp = getTemplate(template);
+			return FreeMarkerTemplateUtils.processTemplateIntoString(temp, params);
+		} catch (Exception e) {
+			log.error("错误信息：{}", e.getMessage());
+			throw new SystemException("S_UnKnow_Error", e.getMessage());
+		}
 	}
 
 	/**
