+++ laokou-common/laokou-common-core/src/test/java/org/laokou/common/core/Base64UtilTest.java
@@ -19,20 +19,29 @@ package org.laokou.common.core;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.laokou.common.core.utils.MapUtil;
+import org.laokou.common.core.utils.Base64Util;
 
-import java.util.Map;
-import java.util.Set;
+import java.nio.charset.StandardCharsets;
 
 /**
  * @author laokou
  */
-class MapTest {
+class Base64UtilTest {
 
 	@Test
-	void testEmptyMap() {
-		Map<String, Set<String>> map = MapUtil.toUriMap(Map.of(), "test");
-		Assertions.assertTrue(MapUtil.isEmpty(map));
+	void testBase64() {
+		String name = "zzzzzzzzzzzccccc\nccccccccccccccccccc231231qweqweqw";
+		String encodeToString = Base64Util.encodeToString(name.getBytes(StandardCharsets.UTF_8));
+		byte[] decode = Base64Util.decode(encodeToString);
+		Assertions.assertArrayEquals(name.getBytes(StandardCharsets.UTF_8), decode);
+	}
+
+	@Test
+	void testMimeBase64() {
+		String name = "zzzzzzzzzzzccccc\nccccccccccccccccccc231231qweqweqw";
+		String encodeToStringOfMime = Base64Util.encodeToStringOfMime(name.getBytes(StandardCharsets.UTF_8));
+		byte[] mimeDecode = Base64Util.decodeOfMime(encodeToStringOfMime);
+		Assertions.assertArrayEquals(name.getBytes(StandardCharsets.UTF_8), mimeDecode);
 	}
 
 }
