+++ laokou-common/laokou-common-core/src/test/java/org/laokou/common/core/TemplateUtilTest.java
@@ -19,29 +19,18 @@ package org.laokou.common.core;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.laokou.common.core.utils.Base64Util;
+import org.laokou.common.core.utils.TemplateUtil;
 
-import java.nio.charset.StandardCharsets;
+import java.util.Map;
 
 /**
  * @author laokou
  */
-class Base64Test {
+class TemplateUtilTest {
 
 	@Test
-	void testBase64() {
-		String name = "zzzzzzzzzzzccccc\nccccccccccccccccccc231231qweqweqw";
-		String encodeToString = Base64Util.encodeToString(name.getBytes(StandardCharsets.UTF_8));
-		byte[] decode = Base64Util.decode(encodeToString);
-		Assertions.assertArrayEquals(name.getBytes(StandardCharsets.UTF_8), decode);
-	}
-
-	@Test
-	void testMimeBase64() {
-		String name = "zzzzzzzzzzzccccc\nccccccccccccccccccc231231qweqweqw";
-		String encodeToStringOfMime = Base64Util.encodeToStringOfMime(name.getBytes(StandardCharsets.UTF_8));
-		byte[] mimeDecode = Base64Util.decodeOfMime(encodeToStringOfMime);
-		Assertions.assertArrayEquals(name.getBytes(StandardCharsets.UTF_8), mimeDecode);
+	void testTemplate() {
+		Assertions.assertEquals("1234", TemplateUtil.getContent("123${code}", Map.of("code","4")));
 	}
 
 }
